# java/fhircast/library/doc/openapi.yml
openapi: '3.0.2'
info:
  title: FHIR Cast
  version: 'STU2 ++'
servers:
  - url: https://api.server.test/v1
paths:
  /sync/fhircast:
    post: 
      summary: subscribe to an event or request a event notification to be send
      description: subscribe to events on the indicated channel.
      parameters: 
      requestBody:
        description: subscription or even data
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/SubscriptionRequest"
                - $ref: "#/components/schemas/ContextChangeRequest"

      responses:
        202: 
          description: successfully subscribed
          content:
            application/json:
              schema:
                type: string
                description: hub_challenge
        200: 
          description: successfully subscribed
          content:
            application/json:
              schema:
                type: string
                description: hub_challenge
  /sync/topic:
    get:
      summary: retrieve available topics
      description: Returns the topics open to current user
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  topic:
                    type: array
                    items:
                      description: array with topics that can be accessed by the current user.
                      type: object
                      properties:
                        topicId:
                          type: string
                          description: topicId of the open topic.
                        hubUrl:
                          type: string
                          description: the url of the fhircast hub used (/sync/fhircast)
                        scratchpad:
                          type: string
                          description: the URL on which the data in the scratchpad can be retrieved as a FHIR endpoint
                        context:
                          type: string
                          description: the URL on which the data in the scratchpad and FHIR server can be retrieved as a merged FHIR endpoint
        '400':
          description: The topic was not created
        '404':
          description: The end-point does not exist
        '500':
          description: something went wrong
    put:
      summary: create topic
      description: create topic with a specific ID
      parameters:
        - name : topic
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successfully created the topic with the indicated id.
        '400':
          description: The topic was not created
        '404':
          description: The end-point does not exist
        '500':
          description: something went wrong
    post:
      summary: create a new FHIRcast topic
      description: A new topic is created
      responses:
        '200':
          description: a topic has been created
          content:
            application/json:
              schema:
                type: object
                properties:
                  topic:
                    type: string
                    example: 9034oiwjfh498-uf98djh98-sas
        '400':
          description: The topic was not created
        '404':
          description: The end-point does not exist
        '500':
          description: something went wrong
  /sync/topic/{topicId}:
    get:
      summary: retrieve information on the topic
      description: Returns current information related to a topic.
      parameters:
        - name : topic
          in: path
          required: true
          schema:
            type: string
    delete:
      summary: delete topic 'topicId'
      description: delete the topic  'topicId'. A log-out event will be send.
      parameters:
        - name : topic
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successfully deleted the topic
        '400':
          description: The topic was not deleted
        '404':
          description: The end-point does not exist
        '500':
          description: something went wrong

    
components:
  schemas:
    SubscriptionRequest:
      type: object
      description: request a subscription
      properties:
        hub.channel.type: 
          type: string
          description: webhook|websocket
        hub.mode: 
          type: string
          description: subscribe|unsubscribe
        hub.topic: 
          type: string
          description: <topic>
        hub.event: 
          type: string
        hub.lease_seconds: 
          type: string
        hub.callback: 
          type: string
        hub.secret: 
          type: string
        hub.challenge: 
          type: string
        hub.channel.endpoint: 
          type: string
    ContextChangeRequest:
      type: object
      description: request a subscription
      properties:
        timestamp: 
          type: string
        id: 
          type: string
        event:
          type: object
          properties:
            hub.topic: 
              type: string
              description: <topic>
            hub.event: 
              type: string
            context: 
              type: array
              items:
                type: Object
                properties:
                  key: 
                    type: string
                  resource: 
                    type: string